Use fingerprint instead of id for gpg key
GPG short ids are vulnerable to collisions (eg https://evil32.com/ , http://security.stackexchange.com/questions/74009/what-is-an-openpgp-key-id-collision). I believe it is *much* harder to create a full fingerprint collision.

We can specify the full fingerprint instead of the key id in order to mitigate an attacker that can MiTM the gpg key provided by packagecloud.

I understand that the key is served over SSL. After this patch, our trust in packagecloud is now stronger than our trust in the SSL connection to packagecloud.io.